Code to detect cycles in the acyclic graph.
Special precautions are taken for parameters that refer to eachother in the same node, e.g. add1.v1 - > add1.v2 . Depending on the evaluation order, this might cause errors. Now, the system resolves these dependencies and only processes the node when all parameters have their dependencies resolved.

This system can detect external cycles, but internal cycles are still silently thrown away. The system will not go into infinite recursion, but it will not show an error either. To prevent these types of cycles to happen in the first place, I would need to rewrite the CycleDetector to work on a per-parameter basis. However, the events that can trigger changes for expressions are hard to determine. (For example, renaming a node might cause an expression to change behaviour, causing a cycle) Instead, I might want to detect that I have processed the same parameter twice at runtime, and throw a processing exception then.