Fixed a bug where some expression connections were not restored on file open.
The problem was that the nodes the expression depended on were not created from the XML because they were not encountered yet. Now, we don't set the expressions directly when we encounter them, but we put them in a temporary map and only set them when the network ends (so we know we have all the nodes created). This is still not perfect for expressions that span networks.