Reverted the decision of not cloning node output.
This gave unforseen problems with nodes re-evaluating and causing bad dirty propagation. Now, parameters clone their values when asked for them and nodes respect dirty flags when updating. We still need to be able to set a caching flag. Just make sure when implementing that dirty propagation works differently: the current node can be dirty, but its parents can be clean, which is very confusing.