Added support for parameter stamping.
We added a stamp expression to retrieve a value
from the context. Also, we added utility methods
to mark all upstream nodes that have a stamp
expression set as dirty.

This is not a perfect solution since you can have 
stamp keys that are only resolved in later nodes. 
However, this approach is correct, but will 
evaluate too much nodes.

The goal for stamping is to allow you to write a 
copy node with stamping support. For an example 
of how this might look, check out ExpressionTest 
or NodeTest.