Better geometry resampling.
Path and Contour have support for resampling based
on amount or length. The implementations still have
a bug for paths with amounts over all of the contours,
and the behaviour for a geometric Group is missing.
The resample node was rewritten, and works in all 
but the case for per-path resampling.

As a result, the unit tests fail for this case, in preparation for a fix.
Canvas is no longer an OldGroup, this class was removed. It just contains a
list of Grobs now.

Also, Grobs have support for rotate/translate/scale.