Added support for working with expressions with errors.
You can now store/load scripts with expression errors, and the errors are propagated correctly. I did this by changing the error handling. Setting the expression no longer causes an error, you have to check hasExpressionError(). It tries its best to set expression dependencies and removes them if an error occurs. To make things clearer, you're now required to catch ExpressionErrors.

I also discovered an error with propagation in the Node: if an error happens during dependency updating, this error is not caught but propagated to the parent, which caused the dialog boxes to pop up instead of coloring the node red.