The node now holds the data class instead of the port.
All Ports on a node accept one type of data, that of the node data class. They no longer provide an option to set the data class.

I removed the access to the data class entirely from the port. If you're accessing the data class on the port, you're probably doing something wrong.