Geometry is placed correctly on the canvas when exporting.
The bounding box of the geometry was converted incorrectly when preparing the document for export. I fixed this by taking the bounding box into account, and allowing PDFRenderer direct access to the Grob to render. This means that we don't have to transform the Grob, but can just move the PDF graphics context.

In addition, I made sure that exported files always have the ".pdf" extension.