Initial commit of the Core Image library.
Functionality of the library is limited to a generator and filter node.

Underlying support for image manipulation is not implemented yet. Users need to drop down to Java's AWT image support to manipulate images. This is a stop-gap solution until we provide official support. We'd rather provide *something* than let the user create ugly hacks to work around the system.