Fixed transformation and bounds bugs in SVG library
Some paths (rects, circles and ovals) couldn't be transformed correctly because (0, 0) would be used as the center as opposed to the upper left corner of the shape (which is used in SVG).

The computation of a paths bounds takes into account every point being added. Since the original SVG library added a MOVETO 0,0 for every single path bounds could get huge very quickly in this implementation. Since the point isn't needed it is omitted, causing the correct bounds to be returned.