The result of an expression is now converted with more leniency.
Before, an error was thrown if the resulting type of an expression did not match the expected type exactly. An exception was made for float/int conversion. We now do more type conversions even if those would result in some information loss, e.g. when converting from float to int.