Updated the Windows installer.
The installer includes the Java JRE. It also defines user-specific properties for the Java JRE and NSIS, which can be customized in build.properties.

The installer also has a header image and a sidebar image on the finish page. It creates a NodeBox icon on the desktop and allows you to run NodeBox when the installer has finished.