Added the export flag.
The export flag has an effect on which nodes can be seen as exported from outside of the library. NodeLibrary.get() only returns nodes that have the export flag set. Nodes need to be explicitly marked as "export" to see them in the node manager. The export flag does not get copied when a new instance is created based on an exported prototype. (Otherwise, the rule would be "once exported, always exported", which would make the flag useless)