The viewer layer now has the correct bounds.
Previously, zooming in too far would stop showing the canvas. This came because of incorrect bounds. Now, we set the bounds of the viewer layer to the bounds of the grob. We have a special case for non-grob objects.

The viewer layer is now set using setOffset() instead of moved to the center in the drawing.

This can be further be optimized by not repainting the entire viewer but just the viewer layer. However, for that to work, the handles also need to be drawn inside of the bounds of the grob, which is not the case.