Enlarged the draggable zone for the output port.
Users would often overshoot the node causing it to drag instead of connecting. I enlarged the zone, making it slightly bigger than the output port so that users have more chance to hit it. This should not interfere with dragging, since that mostly happens from the middle and not around the output port.