Removed the "smart connect" functionality.
Smart connect connected a new node to the already selected node if they were available. This caused some confusion with users where the connection would be made to the wrong port automatically (e.g. when creating a place node).

I'm taking out right now, but we'll examine  if it needs to be back in, maybe in a more subtle way.