The viewer no longer requests focus on mouse enter.
In the default layout, the viewer is above the code window. When drag-selecting text in the code pane and moving up to the viewer, the viewer will request focus and steal the focus from the code pane. This will turn the selected text gray and disallow you to copy/paste code.