Fixed a bug where ndbx documents wouldn't load on systems with the Turkish locale.
The problem happened when strings containing "i" were converted to uppercase. The corevector.ndbx file contains numerous "string" types or "int" types. On loading the ndbx these are converted to uppercase first, trying to match the resulting "STRİNG" or "İNT" with the types in the Parameter.Type enum, which aren't valid of course.