Disabled reload icon when code is unchanged.
Previously, users would hit the reload icon to force an update. This caused the source to change, and thus be included in the node file. By disabling the reload icon, we fix part of the problem.

The other problem still remains: how do we detect identical code so that it is not stored in the file?