Creating a new document does not automatically create a new node.
I tried this for a couple of days and was annoyed by it already. The issue now is that the blank slate still shows canvas properties and most of them don't do anything. I've tried alleviating this problem by changing the order so that the first parameter (background color) does something visible.