Fixed some odd FileDialog behaviour when saving.
FileDialog acted a bit flakey when setFile is called. It did set the right directory but it also shows the entire path in the "file field" of the dialog when an entire path is given. With multiple cancel/ok's the previous path was appended to the current path. This commit addresses that issue.