Make transform handle startup a bit faster.
The transform handle imported everything from the NodeBox graphics
package. This included loading the PDF renderer, which in turn loaded
every font of the system. We now selectively import only the classes we
need, resulting in a speedup.

Now, most time is still spent compiling the transform handle's Python
code. I'm not sure how we can make this faster.