Reloading a function library happens "in place".
Before, a previous "version" of a library was remembered whenever it was reloaded, but then, if you would perform an undo step, the previous version would come to the foreground again. Since the program has no control over what happens to the source code of individual libraries (in case of python and clojure anyway) we want the reloading of libraries be independent of undo/redo.