Each subnet will have the results of its rendered child stored in a separate NodeContext.
Previously there was a problem with nodes that apart from their location within the root path are identical to each other (i.e. /subnet1/copy1 and /subnet2/copy1, here the both "copy1"'s have the same inputs and the same name, the only way they differ is that they are in different subnets) Since we work with immutable nodes, to see if they are a key in the values map, they values are compared, so previously it was possible that the output of copy1 in subnet2 would override the output of copy1 in subnet1. Having a separate NodeContext fixes this problem.