Remove the filter node.
The behavior of the filter node and the cull node is nearly identical.
Since we prefer having the filter name available for when we support
higher-order functions, we'll remove it now and replace it with cull.

Upgrade existing files to change all instances of the filter node with
the cull node.

This fixes #242.