NodeBox files with unicode text are loaded correctly.
Before, we read in files using the default FileReader constructor. This
will create a reader using the default platform encoding which differs
between platforms and even inside and outside of the IDE. This commit
introduces its own method  to remove the ambiguity and always read
files in UTF-8.

For more detail, see
http://stackoverflow.com/questions/696626/java-filereader-encoding-issue

This fixes #351.