Remove prototypes from nodecore.
Nodecore is only used to render the network. All data from the
prototypes is copied over into the JSON data, so no prototype
references are needed.

If we add editing, we might want to re-enable prototypes since then we
have to know if a property has changed with regards to its prototype.